sk_rookies_project4/agent/requirements.txt
opencv-python
pillow
boto3
tensorflow
mysql-connector-python